Added the corresponding files
Officially started the project, basically just trying to apply and learn as I go on. Today was very simple, I gave the basic backbone of the page, and I was trying to implement a navigation bar but I ran into some difficulties. I want to get the layout/formatting of the page perfect before I start adding any raw text since that is a quick fix.